Migrate reusable pieces to a utils submodule.
Remove `make debug` because fathom is no longer runnable. The runnable stuff migrated to tests. Run `make debugtest` to debug it.

Name test files after the modules they test.
