+++ package.json
@@ -24,7 +24,6 @@
     "ranking"
   ],
   "license": "MPL-2.0",
-  "main": "fathom.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/mozilla/fathom.git"
